+++ src/types.ts
@@ -2,11 +2,14 @@
 // vim: set sw=2 ts=2 sts=2 et foldmarker={{,}} foldmethod=marker foldlevel=0 nofen:
 import { Buffer, Neovim, Window } from '@chemzqm/neovim'
 import { CancellationToken, CodeAction, CodeActionKind, CompletionItemLabelDetails, CreateFile, CreateFileOptions, DeleteFile, DeleteFileOptions, Disposable, DocumentSelector, Event, FormattingOptions, Location, Position, Range, RenameFile, RenameFileOptions, SymbolKind, TextDocumentEdit, TextDocumentSaveReason, TextEdit, WorkspaceEdit, WorkspaceFolder } from 'vscode-languageserver-protocol'
-import { TextDocument } from 'vscode-languageserver-textdocument'
-import { URI } from 'vscode-uri'
-import Configurations from './configuration'
-import Document from './model/document'
-import { ProviderResult, TextDocumentContentProvider } from './provider'
+import type { TextDocument } from 'vscode-languageserver-textdocument'
+import type { URI } from 'vscode-uri'
+import type Configurations from './configuration'
+import type Document from './model/document'
+import type RelativePattern from './model/relativePattern'
+import type { ProviderResult, TextDocumentContentProvider } from './provider'
+
+export type GlobPattern = string | RelativePattern
 
 declare global {
   namespace NodeJS {